backend/requirements.txt
google-adk
google-cloud-speech
pydantic
google-generativeai
pytest-mock
